backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-filter==23.1
django-storages==1.13.2
djangorestframework==3.14.0
drf-spectacular==0.26.1
drfpasswordless==1.5.8
google==3.0.0
google-api-core==2.11.0
google-auth==2.17.3
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
idna==3.4
inflection==0.5.1
joblib==1.2.0
jsonschema==4.17.3
numpy==1.24.2
Pillow==9.5.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
random_username @ git+https://github.com/OleksandrPozden/random_username.git@43cc9989192d88e132799084b0a768796c58d603
requests==2.28.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
threadpoolctl==3.1.0
uritemplate==4.1.1
urllib3==1.26.15
